Refactor getFileSizeAsString to use size_t and ensure cleanup.
Updated getFileSizeAsString to take size_t instead of double for consistency and added proper memory cleanup for returned strings.

+++ rbtree.h
@@ -152,7 +152,7 @@ FileInfo parseFileData(const char *inputLine);
 
 char *add_rbt_extension(const char *filename);
 
-char *getFileSizeAsString(double fileSizeBytes);
+char *getFileSizeAsString(size_t fileSizeBytes);
 
 char *get_filename_from_path(const char *path);
